Added Exception

+++ CustomerCare360/src/main/java/org/example/customercare360/Services/BillingCycleServiceImpl.java
@@ -3,20 +3,19 @@ package org.example.customercare360.Services;
 import org.example.customercare360.DTO.BillingCycleRequest;
 import org.example.customercare360.Entity.BillingCycle;
 import org.example.customercare360.Enums.ServiceType;
+import org.example.customercare360.Exception.BillingCycleNotFoundException;
+import org.example.customercare360.Exception.InvalidAccountException;
 import org.example.customercare360.Repository.BillingCycleRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-public class BillingCycleServiceImpl
-        implements BillingCycleService {
+public class BillingCycleServiceImpl implements BillingCycleService {
 
     private final BillingCycleRepository repository;
 
-    public BillingCycleServiceImpl(
-            BillingCycleRepository repository){
-
+    public BillingCycleServiceImpl(BillingCycleRepository repository){
         this.repository = repository;
     }
 
@@ -29,6 +28,14 @@ public class BillingCycleServiceImpl
         cycle.setServiceType(request.getServiceType());
         cycle.setPeriodStart(request.getPeriodStart());
         cycle.setPeriodEnd(request.getPeriodEnd());
+
+//        if(!serviceAccountRepository.existsById(
+//                request.getAccountId())) {
+//
+//            throw new InvalidAccountException(
+//                    "Account Id does not exist"
+//            );
+//        }
 
         return repository.save(cycle);
     }
@@ -44,7 +51,10 @@ public class BillingCycleServiceImpl
             Integer cycleId,
             BillingCycleRequest request){
 
-        BillingCycle cycle = repository.findById(cycleId).orElseThrow();
+        BillingCycle cycle = repository.findById(cycleId)
+                .orElseThrow(() ->
+                        new BillingCycleNotFoundException(
+                                "Billing Cycle not found with id " + cycleId));
 
         cycle.setServiceType(request.getServiceType());
         cycle.setPeriodStart(request.getPeriodStart());
@@ -58,6 +68,11 @@ public class BillingCycleServiceImpl
     @Override
     public String deleteCycle(
             Integer cycleId){
+
+        if(!repository.existsById(cycleId)) {
+            throw new BillingCycleNotFoundException(
+                    "Billing Cycle not found with id " + cycleId);
+        }
 
         repository.deleteById(cycleId);
